requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
apify>=1.5.0

# Docker SDK
docker==7.0.0

# ===== PRIMARY AI PROVIDERS =====
openai==1.3.7
anthropic==0.8.1

# ===== OPTIONAL AI PROVIDERS =====
cohere==4.37
replicate==0.22.0
groq==0.4.1
mistralai==0.1.6
google-genai>=1.0

# HTTP requests
requests==2.31.0
httpx==0.25.2
anyio==3.6.2
aiohttp==3.9.1

# Logging and monitoring
loguru==0.7.2
rich==13.7.0

# Data processing
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
